refactor config pkg
