+++ pkg/config/config.go
@@ -2,17 +2,8 @@ package config
 
 import (
 	"encoding/json"
-	"fmt"
-	"net/url"
-	"os"
-	"strings"
-	"sync"
-
-	"github.com/robfig/cron/v3"
 	"github.com/rs/zerolog"
 )
-
-const DefaultSchedule = "@every 00h00m10s"
 
 type URL string
 
@@ -52,132 +43,4 @@ var validLogLevels = map[string]bool{
 	zerolog.LevelErrorValue: true,
 	zerolog.LevelFatalValue: true,
 	zerolog.LevelPanicValue: true,
-}
-
-type ConfigManager struct {
-	config                  *Config
-	Token                   string
-	DefaultGroundControlURL string
-	configPath              string
-	mu                      sync.RWMutex
-}
-
-func NewConfigManager(path, token, defaultGroundControlURL string, config *Config) (*ConfigManager, error) {
-	return &ConfigManager{
-		config:                  config,
-		configPath:              path,
-		Token:                   token,
-		DefaultGroundControlURL: defaultGroundControlURL,
-	}, nil
-}
-
-func InitConfigManager(path string) (*ConfigManager, []string, error) {
-	cfg, err := ReadAndReturnConfig(path)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to read config: %w", err)
-	}
-
-	warnings, err := ValidateConfig(cfg)
-	if err != nil {
-		return nil, nil, fmt.Errorf("invalid config: %w", err)
-	}
-
-	token := os.Getenv("TOKEN")
-	defaultGroundControlURL := os.Getenv("GROUND_CONTROL_URL")
-
-	cm, err := NewConfigManager(path, token, defaultGroundControlURL, cfg)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to create manager: %w", err)
-	}
-
-	return cm, warnings, nil
-}
-
-// Reads the config at the given path and loads it in the given config variable
-func ReadAndReturnConfig(path string) (*Config, error) {
-	data, err := os.ReadFile(DefaultConfigPath)
-	if err != nil {
-		return nil, err
-	}
-
-	var cfg Config
-	if err := json.Unmarshal(data, &cfg); err != nil {
-		return nil, err
-	}
-
-	return &cfg, nil
-}
-
-func ValidateConfig(config *Config) ([]string, error) {
-	var warnings []string
-
-	if _, err := url.ParseRequestURI(string(config.AppConfig.GroundControlURL)); err != nil {
-		return nil, fmt.Errorf("invalid URL provided for ground_control_url: %w", err)
-	}
-
-	if len(config.ZotConfigRaw) == 0 {
-		return nil, fmt.Errorf("invalid zot_config. zot_config cannot be empty")
-	}
-
-	if config.AppConfig.LogLevel != "" && !validLogLevels[strings.ToLower(config.AppConfig.LogLevel)] {
-		config.AppConfig.LogLevel = zerolog.LevelInfoValue
-		warnings = append(warnings, fmt.Sprintf("invalid log_level '%s' provided. Valid options are: info, debug, panic, error, warn, fatal. Defaulting to 'info'.", config.AppConfig.LogLevel))
-	}
-
-	if !isValidCronExpression(config.AppConfig.StateReplicationInterval) {
-		config.AppConfig.StateReplicationInterval = DefaultSchedule
-		warnings = append(warnings, fmt.Sprintf("invalid schedule provided for StateReplicationInterval, using default schedule %s", DefaultSchedule))
-	}
-
-	if !isValidCronExpression(config.AppConfig.RegisterSatelliteInterval) {
-		config.AppConfig.RegisterSatelliteInterval = DefaultSchedule
-		warnings = append(warnings, fmt.Sprintf("invalid schedule provided for StateReplicationInterval, using default schedule %s", DefaultSchedule))
-	}
-
-	if !isValidCronExpression(config.AppConfig.UpdateConfigInterval) {
-		config.AppConfig.UpdateConfigInterval = DefaultSchedule
-		warnings = append(warnings, fmt.Sprintf("invalid schedule provided for StateReplicationInterval, using default schedule %s", DefaultSchedule))
-	}
-
-	return warnings, nil
-}
-
-func (cm *ConfigManager) WriteConfig() error {
-	cm.mu.Lock()
-	defer cm.mu.Unlock()
-
-	data, err := json.MarshalIndent(cm.config, "", "  ")
-	if err != nil {
-		return err
-	}
-
-	err = os.WriteFile(cm.configPath, data, 0644)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// validateCronExpression checks the validity of a cron expression.
-func isValidCronExpression(cronExpression string) bool {
-	if _, err := cron.ParseStandard(cronExpression); err != nil {
-		return false
-	}
-	return true
-}
-
-func (cm *ConfigManager) With(mutators ...func(*Config)) *ConfigManager {
-	cm.mu.Lock()
-	defer cm.mu.Unlock()
-	for _, mutate := range mutators {
-		mutate(cm.config)
-	}
-	return cm
-}
-
-func (cm *ConfigManager) IsZTRDone() bool {
-	cm.mu.RLock()
-	defer cm.mu.RUnlock()
-	return cm.GetSourceRegistryUsername() != ""
 }
