+++ pkg/config/constants.go
@@ -14,8 +14,8 @@ const DefaultConfigPath string = "config.json"
 
 // Below are the default values of the job schedules that would be used if the user does not provide any schedule or
 // if there is any error while parsing the cron expression
-const DefaultFetchConfigFromGroundControlTimePeriod string = "@every 00h00m30s"
-const DefaultZeroTouchRegistrationCronExpr string = "@every 00h00m05s"
-const DefaultFetchAndReplicateStateTimePeriod string = "@every 00h00m10s"
+const DefaultFetchConfigCronExpr string = "@every 00h00m30s"
+const DefaultZTRCronExpr string = "@every 00h00m05s"
+const DefaultFetchAndReplicateCronExpr string = "@every 00h05m00s"
 
 const BringOwnRegistry bool = false
